% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{selma_funded_statuses}
\alias{selma_funded_statuses}
\alias{SELMA_FUNDED_STATUSES}
\alias{SELMA_ALL_FUNDED_STATUSES}
\title{SELMA Funded Status Groups}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 5.
}
\usage{
SELMA_FUNDED_STATUSES

SELMA_ALL_FUNDED_STATUSES
}
\description{
Groups of status codes representing funded enrolments.
}
\keyword{datasets}
