% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SELMA_FUNDING_LABELS}
\alias{SELMA_FUNDING_LABELS}
\title{Funding source code to label mapping}
\format{
An object of class \code{character} of length 5.
}
\usage{
SELMA_FUNDING_LABELS
}
\description{
Named character vector mapping funding source codes to human-readable labels.
}
\keyword{datasets}
