% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{selma_funding_sources}
\alias{selma_funding_sources}
\alias{SELMA_FUNDING_GOVT}
\alias{SELMA_FUNDING_INTL}
\alias{SELMA_FUNDING_MPTT}
\alias{SELMA_FUNDING_YG}
\alias{SELMA_FUNDING_DQ37}
\title{SELMA Funding Source Codes}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
SELMA_FUNDING_GOVT

SELMA_FUNDING_INTL

SELMA_FUNDING_MPTT

SELMA_FUNDING_YG

SELMA_FUNDING_DQ37
}
\description{
Funding source identifiers used in component-level data.
}
\keyword{datasets}
