% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{selma_join_addresses}
\alias{selma_join_addresses}
\title{Join addresses to students}
\usage{
selma_join_addresses(addresses, students)
}
\arguments{
\item{addresses}{A tibble from \code{\link[=selma_addresses]{selma_addresses()}}.}

\item{students}{A tibble from \code{\link[=selma_students]{selma_students()}}.}
}
\value{
A tibble with address and student columns joined.
}
\description{
Links address records to their associated students.
}
\examples{
\dontrun{
con <- selma_connect()
student_addresses <- selma_join_addresses(
  selma_addresses(con), selma_students(con)
)
}
}
