% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{selma_join_classes}
\alias{selma_join_classes}
\title{Join classes to campuses}
\usage{
selma_join_classes(classes, campuses)
}
\arguments{
\item{classes}{A tibble from \code{\link[=selma_classes]{selma_classes()}}.}

\item{campuses}{A tibble from \code{\link[=selma_campuses]{selma_campuses()}}.}
}
\value{
A tibble with class and campus columns joined.
}
\description{
Links class records to their campus locations. Works on v2 and v3 data.
}
\examples{
\dontrun{
con <- selma_connect()
class_campuses <- selma_join_classes(
  selma_classes(con), selma_campuses(con)
)
}
}
