% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{selma_status_codes}
\alias{selma_status_codes}
\alias{SELMA_STATUS_CONFIRMED}
\alias{SELMA_STATUS_COMPLETED}
\alias{SELMA_STATUS_INCOMPLETE}
\alias{SELMA_STATUS_WITHDRAWN}
\alias{SELMA_STATUS_WITHDRAWN_SDR}
\alias{SELMA_STATUS_EARLY_WITHDRAWN}
\alias{SELMA_STATUS_DEFERRED}
\alias{SELMA_STATUS_CANCELLED}
\alias{SELMA_STATUS_PENDING}
\title{SELMA Enrolment Status Codes}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
SELMA_STATUS_CONFIRMED

SELMA_STATUS_COMPLETED

SELMA_STATUS_INCOMPLETE

SELMA_STATUS_WITHDRAWN

SELMA_STATUS_WITHDRAWN_SDR

SELMA_STATUS_EARLY_WITHDRAWN

SELMA_STATUS_DEFERRED

SELMA_STATUS_CANCELLED

SELMA_STATUS_PENDING
}
\description{
Single-character codes used in enrolment status fields.
}
\keyword{datasets}
