% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selmaR-package.R
\docType{package}
\name{selmaR-package}
\alias{selmaR}
\alias{selmaR-package}
\title{selmaR: R Client for the SELMA Student Management System API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a clean interface to the SELMA student management system REST API. Handles authentication, Hydra JSON-LD pagination, and data standardisation. Supports fetching students, enrolments, intakes, components, and programmes with optional RDS caching. Includes utilities for NZ/AU phone normalisation, SELMA date parsing, and EFTS pro-rata calculation for New Zealand TEOs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pcstrategyandopsco.github.io/selmaR/}
  \item \url{https://github.com/pcstrategyandopsco/selmaR}
  \item Report bugs at \url{https://github.com/pcstrategyandopsco/selmaR/issues}
}

}
\author{
\strong{Maintainer}: Peeyoosh Chandra \email{pc@strategyandops.co}

}
\keyword{internal}
