% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_phones}
\alias{standardize_phones}
\title{Vectorised phone standardisation}
\usage{
standardize_phones(phones, default_regions = c("NZ", "AU"))
}
\arguments{
\item{phones}{Character vector of phone numbers.}

\item{default_regions}{Character vector of region codes to try
(default \code{c("NZ", "AU")}).}
}
\value{
Character vector of standardised E.164 phone numbers.
}
\description{
Applies \code{\link[=standardize_phone]{standardize_phone()}} to each element of a character vector.
}
\examples{
standardize_phones(c("021 123 4567", "+61412345678", NA))
}
