% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{selma_join_notes}
\alias{selma_join_notes}
\title{Join notes to students}
\usage{
selma_join_notes(notes, students)
}
\arguments{
\item{notes}{A tibble from \code{\link[=selma_notes]{selma_notes()}}.}

\item{students}{A tibble from \code{\link[=selma_students]{selma_students()}}.}
}
\value{
A tibble with note and student columns joined.
}
\description{
Links notes/events to student records.
}
\examples{
\dontrun{
con <- selma_connect()
notes_with_students <- selma_join_notes(
  selma_notes(con), selma_students(con)
)
}
}
